renames rootSaga file to groupsSagas and renames the dir containing all sagas related to groups to groups/
